pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dqar"
version = "0.1.0"
description = "Dynamic and Quantization-Aware Attention Reuse for Diffusion Transformers"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Gautham Satyanarayana"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "einops>=0.7.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "clean-fid>=0.1.35",
    "clip-benchmark>=1.4.0",
    "open-clip-torch>=2.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
dqar-train = "scripts.train_policy:main"
dqar-infer = "scripts.run_inference:main"
dqar-eval = "scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dqar*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
